requirements.txt
pycrypto==2.6.1
requests==2.21.0
urllib3==1.24.1
colorlog==4.0.2
pyyaml==3.13
bitmath==1.3.3.1
pyOpenSSL==19.0.0
dnspython==1.16.0
pythonwhois==2.4.3
shodan==1.11.0
scandir==1.9.0
pyravendb==4.0.6.0